Fix tests

+++ asreview/webapp/_api/projects.py
@@ -902,24 +902,29 @@ def api_import_project():
         "settings_metadata.json",
     )
 
-    warnings = []
-    try:
-        ActiveLearningCycle.from_file(fp_al_cycle)
-    except ValueError as err:
-        with open(fp_al_cycle, "w") as f:
-            model = get_ai_config()
-            json.dump(
-                {"name": model["name"], "current_value": asdict(model["value"])}, f
-            )
+    if project.config["mode"] != PROJECT_MODE_SIMULATE:
+        with open(fp_al_cycle, "r") as f:
+            current_cycle = json.load(f)["current_value"]
 
-        warnings.append(
-            str(err) + " It might be removed from this version of ASReview LAB or you "
-            "need to install an extension to use this model component."
-        )
-        warnings.append(
-            " The active learning model has been reset to the default model and"
-            " can be changed in the project settings."
-        )
+        warnings = []
+        try:
+            ActiveLearningCycle.from_meta(ActiveLearningCycleData(**current_cycle))
+        except ValueError as err:
+            with open(fp_al_cycle, "w") as f:
+                model = get_ai_config()
+                json.dump(
+                    {"name": model["name"], "current_value": asdict(model["value"])}, f
+                )
+
+            warnings.append(
+                str(err)
+                + " It might be removed from this version of ASReview LAB or you "
+                "need to install an extension to use this model component."
+            )
+            warnings.append(
+                " The active learning model has been reset to the default model and"
+                " can be changed in the project settings."
+            )
 
     if current_app.config.get("AUTHENTICATION", True):
         current_user.projects.append(Project(project_id=project.config.get("id")))
